♻️ refactor(sharedUI): extract contributor fold + Full Cast overlay to a shared package
