+++ sharedUI/src/commonMain/kotlin/com/calypsan/listenup/client/features/bookdetail/components/BookDetailHero.kt
@@ -26,7 +26,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
@@ -39,13 +38,14 @@ import com.calypsan.listenup.client.design.theme.DisplayFontFamily
 import com.calypsan.listenup.client.design.theme.Spacing
 import com.calypsan.listenup.client.domain.model.BookContributor
 import com.calypsan.listenup.client.domain.model.BookSeries
+import com.calypsan.listenup.client.features.contributors.CastRole
+import com.calypsan.listenup.client.features.contributors.ClickableContributorLine
 import com.calypsan.listenup.client.presentation.bookdetail.HERO_CONTRIBUTOR_FOLD_LIMIT
 import listenup.composeapp.generated.resources.Res
 import listenup.composeapp.generated.resources.book_detail_narrated_by
 import listenup.composeapp.generated.resources.book_detail_other_authors
 import listenup.composeapp.generated.resources.book_detail_other_narrators
 import listenup.composeapp.generated.resources.series_book_sequence
-import org.jetbrains.compose.resources.StringResource
 import org.jetbrains.compose.resources.stringResource
 
 /**
@@ -200,90 +200,6 @@ fun CompactHero(
                 overflowTextRes = Res.string.book_detail_other_narrators,
                 onOverflowClick = { onShowCast(CastRole.Narrators) },
             )
-        }
-    }
-}
-
-/**
- * A centered line of contributor names where each name is individually tappable, joined by
- * ", " / " & " separators — optionally preceded by a [leadingIcon] and a [prefix] string
- * (e.g. the narrator line's microphone icon and "Narrated by ").
- *
- * @param contributors Names to render; the caller guarantees a non-empty list
- * @param onContributorClick Invoked with the tapped contributor's id
- * @param style Text style applied to names, separators, and prefix
- * @param nameColor Colour for the tappable names
- * @param separatorColor Colour for separators, the prefix, and the leading icon's neighbours
- * @param leadingIcon Optional icon rendered before the prefix
- * @param prefix Optional non-tappable text rendered before the first name
- * @param foldLimit Show at most this many names inline; beyond it the line folds to a single
- *   "{lead}, N other …" summary. Folding only happens when both [overflowTextRes] and
- *   [onOverflowClick] are supplied; defaults to [Int.MAX_VALUE] (never fold).
- * @param overflowTextRes Format string for the folded summary, taking the lead name (`%1$s`) and
- *   the other-count (`%2$d`) — e.g. "%1$s, %2$d other narrators".
- * @param onOverflowClick Invoked when the folded line is tapped; wired to open the full-cast overlay
- */
-@OptIn(ExperimentalLayoutApi::class)
-@Composable
-private fun ClickableContributorLine(
-    contributors: List<BookContributor>,
-    onContributorClick: (contributorId: String) -> Unit,
-    style: TextStyle,
-    nameColor: Color,
-    separatorColor: Color,
-    modifier: Modifier = Modifier,
-    horizontalArrangement: Arrangement.Horizontal = Arrangement.spacedBy(0.dp, Alignment.CenterHorizontally),
-    leadingIcon: (@Composable () -> Unit)? = null,
-    prefix: String? = null,
-    foldLimit: Int = Int.MAX_VALUE,
-    overflowTextRes: StringResource? = null,
-    onOverflowClick: (() -> Unit)? = null,
-) {
-    // Fold to "{lead}, N other …" once the inline list would exceed [foldLimit]; the whole line
-    // becomes one tap target that opens the full-cast overlay.
-    val folded = overflowTextRes != null && onOverflowClick != null && contributors.size > foldLimit
-
-    FlowRow(
-        modifier = if (folded) modifier.clickable(onClick = onOverflowClick) else modifier,
-        horizontalArrangement = horizontalArrangement,
-        verticalArrangement = Arrangement.Center,
-    ) {
-        if (leadingIcon != null) {
-            Box(
-                modifier = Modifier.padding(end = 6.dp),
-                contentAlignment = Alignment.Center,
-            ) {
-                leadingIcon()
-            }
-        }
-
-        if (prefix != null) {
-            Text(text = prefix, style = style, color = separatorColor)
-        }
-
-        if (folded) {
-            // Lead and "N others" share one colour — secondary names are not made more prominent
-            // than the lead (the whole line is the affordance into the full cast).
-            Text(
-                text = stringResource(overflowTextRes, contributors.first().name, contributors.size - 1),
-                style = style,
-                color = nameColor,
-            )
-            return@FlowRow
-        }
-
-        contributors.forEachIndexed { index, contributor ->
-            Text(
-                text = contributor.name,
-                style = style,
-                color = nameColor,
-                modifier = Modifier.clickable { onContributorClick(contributor.id) },
-            )
-
-            if (index < contributors.lastIndex) {
-                val separator = if (index == contributors.size - 2) " & " else ", "
-                Text(text = separator, style = style, color = separatorColor)
-            }
         }
     }
 }
